pyproject.toml
[project]
name = "cto4ai-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cairosvg>=2.8.2",
    "pillow>=11.3.0",
    "requests>=2.32.5",
]
